long policy
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@5846 a1c6a512-1295-4272-9138-f99709370657

+++ uisimulator/win32/button.c
@@ -216,7 +216,7 @@ void button_event(int key, bool pressed)
 
 /* Again copied from real button.c... */
 
-int button_get(bool block)
+long button_get(bool block)
 {
     struct event ev;
 
@@ -227,7 +227,7 @@ int button_get(bool block)
     return BUTTON_NONE;
 }
 
-int button_get_w_tmo(int ticks)
+long button_get_w_tmo(int ticks)
 {
     struct event ev;
     queue_wait_w_tmo(&button_queue, &ev, ticks);
